bigint

+++ packages/mountainsort2/src/p_whiten.cpp
@@ -14,9 +14,9 @@ double quantize(float X, double unit)
     return (floor(X / unit + 0.5)) * unit;
 }
 
-void quantize(int N, float* X, double unit)
+void quantize(bigint N, float* X, double unit)
 {
-    for (int i = 0; i < N; i++) {
+    for (bigint i = 0; i < N; i++) {
         X[i] = quantize(X[i], unit);
     }
 }
@@ -27,14 +27,14 @@ bool p_whiten(QString timeseries, QString timeseries_out, Whiten_opts opts)
     (void)opts;
 
     DiskReadMda32 X(timeseries);
-    int M = X.N1();
-    int N = X.N2();
+    bigint M = X.N1();
+    bigint N = X.N2();
 
-    int processing_chunk_size = 1e7;
+    bigint processing_chunk_size = 1e7;
 
     Mda XXt(M, M);
     double* XXtptr = XXt.dataPtr();
-    int chunk_size = processing_chunk_size;
+    bigint chunk_size = processing_chunk_size;
     if (N < processing_chunk_size) {
         chunk_size = N;
     }
@@ -42,9 +42,9 @@ bool p_whiten(QString timeseries, QString timeseries_out, Whiten_opts opts)
     {
         QTime timer;
         timer.start();
-        int num_timepoints_handled = 0;
+        bigint num_timepoints_handled = 0;
 #pragma omp parallel for
-        for (int timepoint = 0; timepoint < N; timepoint += chunk_size) {
+        for (bigint timepoint = 0; timepoint < N; timepoint += chunk_size) {
             Mda32 chunk;
 #pragma omp critical(lock1)
             {
@@ -53,11 +53,11 @@ bool p_whiten(QString timeseries, QString timeseries_out, Whiten_opts opts)
             float* chunkptr = chunk.dataPtr();
             Mda XXt0(M, M);
             double* XXt0ptr = XXt0.dataPtr();
-            for (int i = 0; i < chunk.N2(); i++) {
-                int aa = M * i;
-                int bb = 0;
-                for (int m1 = 0; m1 < M; m1++) {
-                    for (int m2 = 0; m2 < M; m2++) {
+            for (bigint i = 0; i < chunk.N2(); i++) {
+                bigint aa = M * i;
+                bigint bb = 0;
+                for (bigint m1 = 0; m1 < M; m1++) {
+                    for (bigint m2 = 0; m2 < M; m2++) {
                         XXt0ptr[bb] += chunkptr[aa + m1] * chunkptr[aa + m2];
                         bb++;
                     }
@@ -65,23 +65,23 @@ bool p_whiten(QString timeseries, QString timeseries_out, Whiten_opts opts)
             }
 #pragma omp critical(lock2)
             {
-                int bb = 0;
-                for (int m1 = 0; m1 < M; m1++) {
-                    for (int m2 = 0; m2 < M; m2++) {
+                bigint bb = 0;
+                for (bigint m1 = 0; m1 < M; m1++) {
+                    for (bigint m2 = 0; m2 < M; m2++) {
                         XXtptr[bb] += XXt0ptr[bb];
                         bb++;
                     }
                 }
                 num_timepoints_handled += qMin(chunk_size, N - timepoint);
                 if ((timer.elapsed() > 5000) || (num_timepoints_handled == N)) {
-                    printf("%d/%d (%d%%)\n", num_timepoints_handled, N, (int)(num_timepoints_handled * 1.0 / N * 100));
+                    printf("%ld/%ld (%d%%)\n", num_timepoints_handled, N, (int)(num_timepoints_handled * 1.0 / N * 100));
                     timer.restart();
                 }
             }
         }
     }
     if (N > 1) {
-        for (int ii = 0; ii < M * M; ii++) {
+        for (bigint ii = 0; ii < M * M; ii++) {
             XXtptr[ii] /= (N - 1);
         }
     }
@@ -96,9 +96,9 @@ bool p_whiten(QString timeseries, QString timeseries_out, Whiten_opts opts)
     {
         QTime timer;
         timer.start();
-        int num_timepoints_handled = 0;
+        bigint num_timepoints_handled = 0;
 #pragma omp parallel for
-        for (int timepoint = 0; timepoint < N; timepoint += chunk_size) {
+        for (bigint timepoint = 0; timepoint < N; timepoint += chunk_size) {
             Mda32 chunk_in;
 #pragma omp critical(lock1)
             {
@@ -107,11 +107,11 @@ bool p_whiten(QString timeseries, QString timeseries_out, Whiten_opts opts)
             float* chunk_in_ptr = chunk_in.dataPtr();
             Mda32 chunk_out(M, chunk_in.N2());
             float* chunk_out_ptr = chunk_out.dataPtr();
-            for (int i = 0; i < chunk_in.N2(); i++) { // explicitly do mat-mat mult ... TODO replace w/ BLAS3
-                int aa = M * i;
-                int bb = 0;
-                for (int m1 = 0; m1 < M; m1++) {
-                    for (int m2 = 0; m2 < M; m2++) {
+            for (bigint i = 0; i < chunk_in.N2(); i++) { // explicitly do mat-mat mult ... TODO replace w/ BLAS3
+                bigint aa = M * i;
+                bigint bb = 0;
+                for (bigint m1 = 0; m1 < M; m1++) {
+                    for (bigint m2 = 0; m2 < M; m2++) {
                         chunk_out_ptr[aa + m1] += chunk_in_ptr[aa + m2] * WWptr[bb]; // actually this does dgemm w/ WW^T
                         bb++; // but since symmetric, doesn't matter.
                     }
@@ -125,7 +125,7 @@ bool p_whiten(QString timeseries, QString timeseries_out, Whiten_opts opts)
                 Y.writeChunk(chunk_out, 0, timepoint);
                 num_timepoints_handled += qMin(chunk_size, N - timepoint);
                 if ((timer.elapsed() > 5000) || (num_timepoints_handled == N)) {
-                    printf("%d/%d (%d%%)\n", num_timepoints_handled, N, (int)(num_timepoints_handled * 1.0 / N * 100));
+                    printf("%ld/%ld (%d%%)\n", num_timepoints_handled, N, (int)(num_timepoints_handled * 1.0 / N * 100));
                     timer.restart();
                 }
             }
@@ -141,7 +141,7 @@ bool p_compute_whitening_matrix(QStringList timeseries_list, QString whitening_m
     (void)opts;
 
     DiskReadMda32 X0(2, timeseries_list);
-    int M = X0.N1();
+    bigint M = X0.N1();
     bigint N = X0.N2();
     qDebug() << "M/N" << M << N;
 
@@ -164,12 +164,12 @@ bool p_compute_whitening_matrix(QStringList timeseries_list, QString whitening_m
             chunks << chunk0;
             timepoint += chunk_size;
         }
-        int num_chunks = chunks.count();
+        bigint num_chunks = chunks.count();
         qDebug() << QString("Processing %1 chunks. timepoint %2. (%3%)").arg(num_chunks).arg(timepoint).arg((int)(timepoint * 1.0 / N * 100));
 #pragma omp parallel
         {
 #pragma omp for
-            for (int i = 0; i < num_chunks; i++) {
+            for (bigint i = 0; i < num_chunks; i++) {
                 Mda32 chunk0;
 #pragma omp critical
                 {
@@ -180,9 +180,9 @@ bool p_compute_whitening_matrix(QStringList timeseries_list, QString whitening_m
                 float* chunkptr = chunk0.dataPtr();
                 for (bigint i = 0; i < chunk0.N2(); i++) {
                     bigint aa = M * i;
-                    int bb = 0;
-                    for (int m1 = 0; m1 < M; m1++) {
-                        for (int m2 = 0; m2 < M; m2++) {
+                    bigint bb = 0;
+                    for (bigint m1 = 0; m1 < M; m1++) {
+                        for (bigint m2 = 0; m2 < M; m2++) {
                             XXt0ptr[bb] += chunkptr[aa + m1] * chunkptr[aa + m2];
                             bb++;
                         }
@@ -190,9 +190,9 @@ bool p_compute_whitening_matrix(QStringList timeseries_list, QString whitening_m
                 }
 #pragma omp critical(lock2)
                 {
-                    int bb = 0;
-                    for (int m1 = 0; m1 < M; m1++) {
-                        for (int m2 = 0; m2 < M; m2++) {
+                    bigint bb = 0;
+                    for (bigint m1 = 0; m1 < M; m1++) {
+                        for (bigint m2 = 0; m2 < M; m2++) {
                             XXtptr[bb] += XXt0ptr[bb];
                             bb++;
                         }
@@ -203,7 +203,7 @@ bool p_compute_whitening_matrix(QStringList timeseries_list, QString whitening_m
     }
 
     if (N > 1) {
-        for (int ii = 0; ii < M * M; ii++) {
+        for (bigint ii = 0; ii < M * M; ii++) {
             XXtptr[ii] /= (N - 1);
         }
     }
@@ -224,19 +224,19 @@ bool p_whiten_clips(QString clips_path, QString whitening_matrix, QString clips_
     double* WWptr = WW.dataPtr();
     float* clips_ptr = clips.dataPtr();
 
-    int M = clips.N1();
-    int T = clips.N2();
-    int L = clips.N3();
+    bigint M = clips.N1();
+    bigint T = clips.N2();
+    bigint L = clips.N3();
 
     Mda32 clips_out(M, T, L);
     float* clips_out_ptr = clips_out.dataPtr();
 
-    for (int i = 0; i < L; i++) {
-        for (int t = 0; t < T; t++) {
-            int aa = M * T * i + M * t;
-            int bb = 0;
-            for (int m1 = 0; m1 < M; m1++) {
-                for (int m2 = 0; m2 < M; m2++) {
+    for (bigint i = 0; i < L; i++) {
+        for (bigint t = 0; t < T; t++) {
+            bigint aa = M * T * i + M * t;
+            bigint bb = 0;
+            for (bigint m1 = 0; m1 < M; m1++) {
+                for (bigint m2 = 0; m2 < M; m2++) {
                     clips_out_ptr[aa + m1] += clips_ptr[aa + m2] * WWptr[bb]; // actually this does dgemm w/ WW^T
                     bb++; // but since symmetric, doesn't matter.
                 }
@@ -258,11 +258,11 @@ bool p_apply_whitening_matrix(QString timeseries, QString whitening_matrix, QStr
     (void)opts;
 
     DiskReadMda32 X(timeseries);
-    int M = X.N1();
-    int N = X.N2();
+    bigint M = X.N1();
+    bigint N = X.N2();
 
-    int processing_chunk_size = 1e7;
-    int chunk_size = processing_chunk_size;
+    bigint processing_chunk_size = 1e7;
+    bigint chunk_size = processing_chunk_size;
     if (N < processing_chunk_size) {
         chunk_size = N;
     }
@@ -276,9 +276,9 @@ bool p_apply_whitening_matrix(QString timeseries, QString whitening_matrix, QStr
     {
         QTime timer;
         timer.start();
-        int num_timepoints_handled = 0;
+        bigint num_timepoints_handled = 0;
 #pragma omp parallel for
-        for (int timepoint = 0; timepoint < N; timepoint += chunk_size) {
+        for (bigint timepoint = 0; timepoint < N; timepoint += chunk_size) {
             Mda32 chunk_in;
 #pragma omp critical(lock1)
             {
@@ -287,11 +287,11 @@ bool p_apply_whitening_matrix(QString timeseries, QString whitening_matrix, QStr
             float* chunk_in_ptr = chunk_in.dataPtr();
             Mda32 chunk_out(M, chunk_in.N2());
             float* chunk_out_ptr = chunk_out.dataPtr();
-            for (int i = 0; i < chunk_in.N2(); i++) { // explicitly do mat-mat mult ... TODO replace w/ BLAS3
-                int aa = M * i;
-                int bb = 0;
-                for (int m1 = 0; m1 < M; m1++) {
-                    for (int m2 = 0; m2 < M; m2++) {
+            for (bigint i = 0; i < chunk_in.N2(); i++) { // explicitly do mat-mat mult ... TODO replace w/ BLAS3
+                bigint aa = M * i;
+                bigint bb = 0;
+                for (bigint m1 = 0; m1 < M; m1++) {
+                    for (bigint m2 = 0; m2 < M; m2++) {
                         chunk_out_ptr[aa + m1] += chunk_in_ptr[aa + m2] * WWptr[bb]; // actually this does dgemm w/ WW^T
                         bb++; // but since symmetric, doesn't matter.
                     }
@@ -305,7 +305,7 @@ bool p_apply_whitening_matrix(QString timeseries, QString whitening_matrix, QStr
                 Y.writeChunk(chunk_out, 0, timepoint);
                 num_timepoints_handled += qMin(chunk_size, N - timepoint);
                 if ((timer.elapsed() > 5000) || (num_timepoints_handled == N)) {
-                    printf("%d/%d (%d%%)\n", num_timepoints_handled, N, (int)(num_timepoints_handled * 1.0 / N * 100));
+                    printf("%ld/%ld (%d%%)\n", num_timepoints_handled, N, (int)(num_timepoints_handled * 1.0 / N * 100));
                     timer.restart();
                 }
             }
